refactor: change to app.service
