+++ packages/backend/src/modules/app-lifecycle/app-lifecycle.service.ts
@@ -499,8 +499,7 @@ export class AppLifecycleService {
   async updateAllApps() {
     const installedApps = await this.appsService.getInstalledApps();
     const availableUpdates = installedApps.filter(
-      ({ app, metadata }) =>
-        Number(app.version) < Number(metadata.latestVersion) && app.ignoredVersion !== metadata.latestVersion,
+      ({ app, metadata }) => Number(app.version) < Number(metadata.latestVersion) && app.ignoredVersion !== metadata.latestVersion,
     );
 
     for (const { app } of availableUpdates) {
@@ -511,34 +510,6 @@ export class AppLifecycleService {
         this.logger.error(`Failed to update app ${app.id}`, e);
       }
     }
-  }
-
-  async ignoreAppVersion(params: { appUrn: AppUrn }) {
-    const { appUrn } = params;
-    const app = await this.appRepository.getAppByUrn(appUrn);
-
-    if (!app) {
-      throw new TranslatableError('APP_NOT_FOUND', { appUrn });
-    }
-
-    const { latestVersion } = await this.marketplaceService.getAppUpdateInfo(appUrn);
-
-    await this.appRepository.updateAppById(app.id, { ignoredVersion: latestVersion });
-
-    this.logger.info(`Ignored version ${latestVersion} for app ${appUrn}`);
-  }
-
-  async unignoreAppVersion(params: { appUrn: AppUrn }) {
-    const { appUrn } = params;
-    const app = await this.appRepository.getAppByUrn(appUrn);
-
-    if (!app) {
-      throw new TranslatableError('APP_NOT_FOUND', { appUrn });
-    }
-
-    await this.appRepository.updateAppById(app.id, { ignoredVersion: null });
-
-    this.logger.info(`Unignored version for app ${appUrn}`);
   }
 
   async startAllApps() {
